Cleanup samples and add dll
