+++ .github/workflows/main.yml
@@ -42,24 +42,26 @@ jobs:
     #  run: soup build ./Soup/Source/Client/CLI/ -v:d -flavor ${{matrix.config}}
 
     # Build Samples
-    - name: Soup Build Cpp ModuleConsoleApplication
-      run: soup build ./Soup/Samples/Cpp/ModuleConsoleApplication/ -flavor ${{matrix.config}}
+    - name: Soup Install Cpp BuildExtension
+      run: soup install ./Soup/Samples/Cpp/BuildExtension/Executable/
+    - name: Soup Build Cpp BuildExtension
+      run: soup build ./Soup/Samples/Cpp/BuildExtension/ -flavor ${{matrix.config}}
+    - name: Soup Build Cpp ConsoleApplication
+      run: soup build ./Soup/Samples/Cpp/ConsoleApplication/ -flavor ${{matrix.config}}
+    - name: Soup Build Cpp DynamicLibrary
+      run: soup build ./Soup/Samples/Cpp/DynamicLibrary/Executable/ -flavor ${{matrix.config}}
+    - name: Soup Build Cpp ModuleInterface
+      run: soup build ./Soup/Samples/Cpp/ModuleInterface/ -flavor ${{matrix.config}}
     - name: Soup Install Cpp ParseJsonFile
       run: soup install ./Soup/Samples/Cpp/ParseJsonFile/
     - name: Soup Build Cpp ParseJsonFile
       run: soup build ./Soup/Samples/Cpp/ParseJsonFile/ -flavor ${{matrix.config}}
-    - name: Soup Install Cpp SimpleBuildExtension
-      run: soup install ./Soup/Samples/Cpp/SimpleBuildExtension/Executable/
-    - name: Soup Build Cpp SimpleBuildExtension
-      run: soup build ./Soup/Samples/Cpp/SimpleBuildExtension/Executable/ -flavor ${{matrix.config}}
-    - name: Soup Build Cpp SimpleConsoleApplication
-      run: soup build ./Soup/Samples/Cpp/SimpleConsoleApplication/ -flavor ${{matrix.config}}
-    - name: Soup Build Cpp StaticLibraryReference
-      run: soup build ./Soup/Samples/Cpp/StaticLibraryReference/MyApplication/ -flavor ${{matrix.config}}
+    - name: Soup Build Cpp StaticLibrary
+      run: soup build ./Soup/Samples/Cpp/StaticLibrary/MyApplication/ -flavor ${{matrix.config}}
 
-    - name: Soup Install CSharp SimpleBuildExtension
-      run: soup install ./Soup/Samples/CSharp/SimpleBuildExtension/Executable/
-    - name: Soup Build CSharp SimpleBuildExtension
-      run: soup build ./Soup/Samples/CSharp/SimpleBuildExtension/Executable/ -flavor ${{matrix.config}}
-    - name: Soup Build CSharp SimpleConsoleApplication
-      run: soup build ./Soup/Samples/CSharp/SimpleConsoleApplication/ -flavor ${{matrix.config}}
+    - name: Soup Install CSharp BuildExtension
+      run: soup install ./Soup/Samples/CSharp/BuildExtension/Executable/
+    - name: Soup Build CSharp BuildExtension
+      run: soup build ./Soup/Samples/CSharp/BuildExtension/Executable/ -flavor ${{matrix.config}}
+    - name: Soup Build CSharp ConsoleApplication
+      run: soup build ./Soup/Samples/CSharp/ConsoleApplication/ -flavor ${{matrix.config}}
