+++ Samples/CSharp/ConsoleApplication/.gitignore
