+++ Samples/CSharp/ConsoleApplication/Recipe.toml
@@ -1,4 +1,4 @@
-Name = "Samples.SimpleConsoleApplication"
+Name = "Samples.CSharp.ConsoleApplication"
 Language = "C#"
 Type = "Executable"
 Version = "1.1.4"
