+++ Samples/Cpp/ConsoleApplication/.gitignore
