+++ Samples/Cpp/ConsoleApplication/Main.cpp
