+++ Samples/Cpp/ConsoleApplication/Recipe.toml
@@ -1,4 +1,4 @@
-Name = "SimpleConsoleApplication"
+Name = "Samples.Cpp.ConsoleApplication"
 Language = "C++"
 Type = "Executable"
 Version = "1.1.3"
