+++ Samples/Cpp/DynamicLibrary/MyApplication/Main.cpp
