+++ Samples/Cpp/DynamicLibrary/MyLibrary/Recipe.toml
@@ -1,0 +1,5 @@
+Name = "Samples.Cpp.DynamicLibrary.MyLibrary"
+Language = "C++"
+Version = "1.0.0"
+Interface = "Module.cpp"
+Type = "DynamicLibrary"
