+++ Samples/Cpp/ModuleInterface/Main.cpp
@@ -1,0 +1,9 @@
+#include <iostream>
+
+import ModuleInterface;
+
+int main()
+{
+	std::cout << "Hello World, " << Helper::GetName() << " Style!" << std::endl;
+	return 0;
+}
