+++ Samples/Cpp/ModuleInterface/Module.cpp
@@ -3,7 +3,7 @@ module;
 // Include all standard library headers in the global module
 #include <string>
 
-export module ModuleConsoleApplication;
+export module ModuleInterface;
 
 export class Helper
 {
