+++ Samples/Cpp/ParseJsonFile/Recipe.toml
@@ -1,4 +1,4 @@
-Name = "ParseJsonFile"
+Name = "Samples.Cpp.ParseJsonFile"
 Language = "C++"
 Version = "1.0.0"
 Type = "Executable"
