+++ Samples/Cpp/StaticLibrary/MyLibrary/Module.cpp
