--- Samples/Cpp/StaticLibraryReference/MyLibrary/Recipe.toml
@@ -1,4 +1,0 @@
-Name = "MyLibrary"
-Language = "C++"
-Version = "1.0.0"
-Interface = "Module.cpp"
